% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{add_time_track}
\alias{add_time_track}
\title{Function adding a time track as a time object, so that it is easier to manipulate}
\usage{
add_time_track(RR, now = "2020-09-05 12:11:00", reverse = FALSE)
}
\arguments{
\item{RR}{RR object}

\item{now}{when does the recording begin? default is "2020-09-05 12:11:00"}

\item{reverse}{whether the RR intervals time series should be reversed (necessary for jumping window)}
}
\value{
data frame with time_track as the first vector
}
\description{
Function adding a time track as a time object, so that it is easier to manipulate
}
