% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{analyze_rr_runs}
\alias{analyze_rr_runs}
\title{Create a new RRRuns object and return the runs analysis from rust}
\usage{
analyze_rr_runs(rr, annotations, write_last_run)
}
\arguments{
\item{rr}{Vector of RR intervals}

\item{annotations}{Vector of annotations (0 for normal beats, non-zero for abnormal)}

\item{write_last_run}{Whether to include the last run in the analysis}
}
\value{
A list containing the runs analysis
}
\description{
Create a new RRRuns object and return the runs analysis from rust
}
