% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{bind_runs_as_table}
\alias{bind_runs_as_table}
\title{Function binding the results of runs calculation for different datasets (files)
into a single table}
\usage{
bind_runs_as_table(results, rownames = NULL)
}
\arguments{
\item{results}{list of results from the `countruns` function}

\item{rownames}{rownames for the table}
}
\value{
data.table
}
\description{
Sometimes many files are calculated and many results objects are collected (one for each dataset / file)
in this situation it may be helpful to present them as a table. This function collects all the results
as rows of a table, padding with zeros when necessary
}
