% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{calculate_frequency_bands}
\alias{calculate_frequency_bands}
\title{Function to calculate power in specific bands}
\usage{
calculate_frequency_bands(spectrum, bands = frequency_bands)
}
\arguments{
\item{spectrum}{list with power and frequency}

\item{bands}{list in the form list(LF=c(0.00, .04)) etc.}
}
\value{
list with bands names as names and power in bands as the values
}
\description{
Function to calculate power in specific bands
}
