% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_runs.R
\name{calculate_label_positions}
\alias{calculate_label_positions}
\title{function drawing a SIMULATED RR runs (for presentation purposes)}
\usage{
calculate_label_positions(runs_vec)
}
\arguments{
\item{runs_vec}{vector with a list of runs names, e.g. c("DR3", "AR4", "N1", "AR3", "DR2")}
}
\value{
list of: simulated RR intervals, label positions, labels
}
\description{
function drawing a SIMULATED RR runs (for presentation purposes)
}
