% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_runs.R
\name{calculate_labels_positions}
\alias{calculate_labels_positions}
\title{function generating the positions of labels to annotate a runs plot}
\usage{
calculate_labels_positions(rr, labels, d = 3, x_offset = 0.4, y_offset = 50)
}
\arguments{
\item{rr}{rr intervals time series}

\item{labels}{vector with a list of runs names, e.g. c("DR3", "AR4", "N1", "AR3", "DR2")}

\item{d}{the distance from segment/point}

\item{x_offset}{offset on the x axis}

\item{y_offset}{offset on the y axis}
}
\value{
list of labels positions in points
}
\description{
function generating the positions of labels to annotate a runs plot
}
