% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{calculate_RR_spectrum}
\alias{calculate_RR_spectrum}
\title{Master function getting the results}
\usage{
calculate_RR_spectrum(RR, bands = frequency_bands)
}
\arguments{
\item{RR}{RR object}

\item{bands}{frequency bands}
}
\value{
list with power in bands
}
\description{
Master function getting the results
}
