% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{check_perp_point}
\alias{check_perp_point}
\title{function checking whether the segments perpendicular to l_1 are not too close from one another}
\usage{
check_perp_point(perp_point, existing_points, centroid_offset, scale)
}
\description{
function checking whether the segments perpendicular to l_1 are not too close from one another
}
