% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{countruns}
\alias{countruns}
\title{Counting monotonic runs in an RR intervals time series.}
\usage{
countruns(rr, annotations = c())
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}
}
\value{
a list with keys dicection_up, direction_down, no_change whose values are vectors containing corresponding runs types statistics
}
\description{
Counts the number of monotonic runs in an RR interval times series, for which
the annotations vector may or may not be available. If the annotations vector
is not available all RRs are assumed to be of sinus origin. This functon
counts all the runs of a specific type (decelerations, accelerations, no
change), up to the maximum values present in the RR intervals time series.
For example, if there is only one deceleration run of the type 1 2 3 4 5, the
result will be direction_up = c(0,0,0,0,1),
direction_down = NULL, no_change = NULL
}
\examples{
countruns(RR$RR, RR$flags)
countruns(RR$RR)

}
\references{
J Piskorski, P Guzik, The structure of heart rate asymmetry: deceleration and acceleration runs, Physiological measurement 32 (8), (2011)
}

