% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{describerr}
\alias{describerr}
\title{RR intervals time series statistics}
\usage{
describerr(rr, annotations)
}
\arguments{
\item{rr}{rr intervals}

\item{annotations}{The annotations in the analyzed RR intervals time series}
}
\value{
an 1 x 6 vector containing the number of all beats (all), sinus beats (N), ventricular beats (V), supraventricular beats (S), artifacts (X) and unidentifed beats (U)
}
\description{
This function returns the statistics of the beats in the analyzed RR
intervals time series
}
