% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{draw_pp_plotly}
\alias{draw_pp_plotly}
\title{Drawing the Poincare plot using base}
\usage{
draw_pp_plotly(PP, vname = "RR", ...)
}
\arguments{
\item{PP}{poincare plot object}

\item{vname}{variable name - this will be used for construction \code{xlab} and \code{ylab} for the Poincare plot}

\item{\\dots}{Additional arguments passed on to \code{plot}}
}
\description{
This function draws the Poincare plot prepared by the \code{preparePP} function
}
