% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{drawpp}
\alias{drawpp}
\title{Drawing the Poincare plot}
\usage{
drawpp(rr, annotations, vname = "RR", ...)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}

\item{vname}{variable name - this will be used for construction \code{xlab} and \code{ylab} for the Poincare plot}

\item{\\dots}{Additional arguments passed on to \code{plot}}
}
\description{
This function draws the Poincare plot prepared by the \code{preparePP} function
}
\examples{
drawpp(RR$RR, RR$flags)
drawpp(RR$RR, RR$flags, vname="PP", col="gray", bg="blue")
}

