% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{drawpp_features}
\alias{drawpp_features}
\title{Drawing reference features of the Poincare plot}
\usage{
drawpp_features(
  rr,
  annotations,
  vname = "RR",
  centroid_offset = 15,
  n_points_1 = 10,
  n_points_2 = 5,
  scale_1 = 0.5,
  scale_2 = 0.4
)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}

\item{vname}{variable name - this will be used for construction \code{xlab} and \code{ylab} for the Poincare plot}

\item{\\dots}{Additional arguments passed on to \code{plot}}
}
\description{
This function draws the Poincare plot in two subplots with the characteristic
object, i.e. the identity line, perpendicular line, centroid and a few select
distances to the characteristic lines
}
\examples{
drawpp_features(RR$RR, RR$flags)
}
