% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{entropies}
\alias{entropies}
\title{Function extracting runs entropy}
\usage{
entropies(nb_decelerations, nb_accelerations, nb_noChanges)
}
\arguments{
\item{nb_decelerations}{vector with the numbers of respective acceleratons}

\item{nb_noChanges}{vector with the numbers of respective neutral_runs}
}
\value{
vector of three folats
}
\description{
Function extracting runs entropy
}
