% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\docType{data}
\name{frequency_bands}
\alias{frequency_bands}
\title{Frequency bands < 24 h}
\format{
An object of class \code{list} of length 3.
}
\usage{
frequency_bands
}
\description{
Frequency bands < 24 h
}
\keyword{datasets}
