% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\docType{data}
\name{frequency_bands_24}
\alias{frequency_bands_24}
\title{Frequency bands 24 h}
\format{
An object of class \code{list} of length 4.
}
\usage{
frequency_bands_24
}
\description{
Frequency bands 24 h
}
\keyword{datasets}
