% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_runs.R
\name{generate_runs_plot}
\alias{generate_runs_plot}
\title{function generating runs plot}
\usage{
generate_runs_plot(
  rr_intervals = NULL,
  labels = c("DR4", "AR4", "N1", "AR3", "DR2")
)
}
\arguments{
\item{rr_intervals}{vector of rr intervals, if null, rr intervals will be simulated on the basis of the provided labels}

\item{labels}{vector of labels like c("DR3", "AR4", "N1", "AR3", "DR2") - if rr_intervals are provided, labels will be ignored}
}
\value{
does not return anything, run for side effects (plotting / exporting plot)
}
\description{
function generating runs plot
}
