% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_longest_run}
\alias{get_longest_run}
\title{Function extracting the longest run from list}
\usage{
get_longest_run(runs_results, type)
}
\arguments{
\item{type}{type of run (up or down)}

\item{runs_result}{list with runs results}
}
\value{
int
}
\description{
Function extracting the longest run from list
}
