% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{get_point_on_line}
\alias{get_point_on_line}
\title{Function returns both x and the coordinate y of a point with coordinate x,
which lies on the line defined by points x1, y1, x2, y2}
\usage{
get_point_on_line(x, x1, y1, x2, y2)
}
\description{
Function returns both x and the coordinate y of a point with coordinate x,
which lies on the line defined by points x1, y1, x2, y2
}
