% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_runs}
\alias{get_runs}
\title{Counting runs in a segment of RR intervals of sinus origin}
\usage{
get_runs(rr, all_runs = list(), directions = c())
}
\arguments{
\item{rr}{vector containing RR intervals of only sinus origin of length at least 2}

\item{all_runs}{list that keeps all the runs in order}

\item{directions}{vector that keeps the designation - whether the run in the all_runs list is a deceleration or aceleration or noChange}
}
\value{
a two-element list, the first element is the list of runs, the second is the vector holding the directions of the runs
}
\description{
This is a helper function used to count runs in a single segment, resulting
from splitting the whole RR time series on non-zero annotations
}
