% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_runs_sequence}
\alias{get_runs_sequence}
\title{Function to get runs sequence in an RR intervals time series, e.g. c("DR3", "AR4", "N1", "AR3", "DR2")}
\usage{
get_runs_sequence(rr, annotations)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}
}
\value{
vector of runs sequence, like c("DR3", "AR4", "N1", "AR3", "DR2")
}
\description{
Function to get runs sequence in an RR intervals time series, e.g. c("DR3", "AR4", "N1", "AR3", "DR2")
}
