% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_runs_summary}
\alias{get_runs_summary}
\title{Get a summary of runs analysis}
\usage{
get_runs_summary(rr, annotations, write_last_run)
}
\arguments{
\item{rr}{Vector of RR intervals}

\item{annotations}{Vector of annotations (0 for normal beats, non-zero for abnormal)}

\item{write_last_run}{Whether to include the last run in the analysis}
}
\value{
A vector of data with rows and columns as attributes
}
\description{
Get a summary of runs analysis
}
