% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{getruns_rust}
\alias{getruns_rust}
\title{Get a summary of runs analysis using rust code}
\usage{
getruns_rust(rr, annotations)
}
\arguments{
\item{rr}{Vector of RR intervals}

\item{annotations}{Vector of annotations (0 for normal beats, non-zero for abnormal)}
}
\value{
a list with results for runs counts and results for runs variances
}
\description{
Get a summary of runs analysis using rust code
}
