% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{index_based_jump}
\alias{index_based_jump}
\title{Function dividing the data.frame into JUMPING windows of `window` length based on index}
\usage{
index_based_jump(RR, window = 300, cut_end = FALSE)
}
\arguments{
\item{RR}{RR data}

\item{window}{the length of the window in beats}

\item{cut_end}{if the window does not fit the entire recording perfectly, should I cut_end = TRUE or the beginning (FALSE) of the recording}

\item{now}{when does the recording begin? default is "2020-09-05 12:11:00"}
}
\description{
Function dividing the data.frame into JUMPING windows of `window` length based on index
}
