% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_runs.R
\name{iseven}
\alias{iseven}
\title{helper function to check if a number is even}
\usage{
iseven(x)
}
\arguments{
\item{x}{number to be checked}
}
\value{
TRUE/FALSE
}
\description{
helper function to check if a number is even
}
