% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{lomb_spectrum}
\alias{lomb_spectrum}
\title{Function calculating Lomb-Scargle periodogram and returning power within the pre-defined bands
power is normalized so that the total power is equal to the variance of the recording}
\usage{
lomb_spectrum(RR)
}
\arguments{
\item{RR}{RR object}
}
\value{
list with power and frequency
}
\description{
Function calculating Lomb-Scargle periodogram and returning power within the pre-defined bands
power is normalized so that the total power is equal to the variance of the recording
}
