% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{name_elements}
\alias{name_elements}
\title{Function giving names to vector elements}
\usage{
name_elements(v, stem)
}
\arguments{
\item{v}{numeric vector}

\item{stem}{stem of the name for vector elements}

\item{return}{numeric vector with named elements}
}
\description{
Function giving names to vector elements
}
