% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen.R
\name{ncm_samp_en}
\alias{ncm_samp_en}
\title{Calculate Sample Entropy Using NCM Algorithm}
\usage{
ncm_samp_en(signal, m, r)
}
\arguments{
\item{signal}{numeric vector, the data for which the Sample Entropy is to be calculated.}

\item{m}{integer, the embedding dimension.}

\item{r}{numeric, the radius of comparison.}
}
\value{
numeric, the calculated sample entropy.
}
\description{
This function calculates the sample entropy for a given signal,
embedding dimension m, and radius of comparison r,
using the NCM algorithm developed by Zurek et al.
}
