% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{parall_line}
\alias{parall_line}
\title{function which returns the Poincare plot parallel line}
\usage{
parall_line(centroid, centroid_offset, rr)
}
\arguments{
\item{centroid}{the x coordinate of the centroid}

\item{centroid_offset}{how far the centroid should be from the paralell line}

\item{rr}{vector with RR intervals}
}
\description{
function which returns the Poincare plot parallel line
}
