% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{perp_line}
\alias{perp_line}
\title{function which returns the Poincare plot perpendicular line}
\usage{
perp_line(centroid, rr, centroid_offset)
}
\arguments{
\item{centroid}{the x coordinate of the centroid}

\item{rr}{vector with RR intervals}

\item{centroid_offset}{how far the centroid should be from the perpendicular line}
}
\description{
function which returns the Poincare plot perpendicular line
}
