% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpp_features.R
\name{perp_point}
\alias{perp_point}
\title{function which returns the coordinates of the projection of a point on l2}
\usage{
perp_point(x, y, centroid_offset)
}
\arguments{
\item{centroid_offset}{centroid offset}

\item{(x, }{y) coordinates of the point to be projected}
}
\description{
function which returns the coordinates of the projection of a point on l2
}
