% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{pnn_perc}
\alias{pnn_perc}
\title{Function calculating pnn%}
\usage{
pnn_perc(pp, threshold, asym = FALSE)
}
\arguments{
\item{pp}{Poincare plot object}

\item{threshold}{threshold for calculating pnnX}

\item{asym}{whether to calculate the asymmetric pnnX}

\item{dec}{whether in asymmetric pnnX we count decelerations to be bigger, false means accelerations are expected to be smaller}
}
\description{
Function calculating pnn%
}
