% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{pp}
\alias{pp}
\title{Wrapper for preparepp - accepts dataframes (with the intention to make this a constructor)}
\usage{
pp(rr_df)
}
\arguments{
\item{rr_df}{dataframe with RR intervals (first column) and annotations (second column)}
}
\value{
poincare plot object
}
\description{
Wrapper for preparepp - accepts dataframes (with the intention to make this a constructor)
}
