% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{preparepp}
\alias{preparepp}
\title{Preparing data for Poincare plot descriptors calculations}
\usage{
preparepp(rr, annotations = c(), throwError = FALSE)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}

\item{throwError}{whether to throw error or return a vector with NAs}
}
\value{
a n x 2 matrix containing the filtered Poincare plot data
}
\description{
This function prepares data for Poincare plot calculations,
namely it prepares the auxiliary rr_i and rr_ii vectors and
filters them appropriately, so that no non-physiological pairs
occur in the Poincare plot
}
