% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{remove_zeros}
\alias{remove_zeros}
\title{Helper function removing zeros resulting from using the slider package}
\usage{
remove_zeros(resulting_windows)
}
\arguments{
\item{resulting_windows}{list of windows}
}
\description{
Helper function removing zeros resulting from using the slider package
}
