% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{reverse_df}
\alias{reverse_df}
\title{Helper function reversing the row order of a data frame}
\usage{
reverse_df(df_elem)
}
\arguments{
\item{df_elem}{the data frame to be reversed}
}
\value{
reversed dataframe
}
\description{
Helper function reversing the row order of a data frame
}
