% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{reverse_time_track}
\alias{reverse_time_track}
\title{Function reversing time track when it is in a list of reversed RR intervals time series}
\usage{
reverse_time_track(RR_time_track, cut_end = FALSE)
}
\description{
Function reversing time track when it is in a list of reversed RR intervals time series
}
