% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RR}
\alias{RR}
\title{RR intervals time series with annotations (flags)}
\format{Dataframe containing an RR intervals time series with annotations.
\describe{
\item{RR}{RR intervals}
\item{flags}{annotations for the RR intervals, 0 stands for a beat of sinus origin}
}}
\source{
my own ECG
}
\usage{
RR
}
\description{
RR intervals time series with annotations (flags)
}
\examples{
\dontrun{
RR
}
}
\keyword{datasets}

