% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{samp_en}
\alias{samp_en}
\title{Get the sample entropy for a signal}
\usage{
samp_en(signal, m, r)
}
\arguments{
\item{signal}{signal for which the sample entropy will be calculated}

\item{m}{embedding dimension - int}

\item{r}{comparison radius - float}
}
\value{
a float
}
\description{
Get the sample entropy for a signal
}
