% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{shave_nonsinus}
\alias{shave_nonsinus}
\title{Function removing artefacts from beginning and end of the recording}
\usage{
shave_nonsinus(RR)
}
\arguments{
\item{RR}{RR object}
}
\description{
Function removing artefacts from beginning and end of the recording
}
