% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_runs.R
\name{simulate_rr_intervals}
\alias{simulate_rr_intervals}
\title{function generating a SIMULATED rr intervals time series}
\usage{
simulate_rr_intervals(runs_vec, starting_rr = c(894))
}
\arguments{
\item{runs_vec}{runs_vec vector with a list of runs names, e.g. c("DR3", "AR4", "N1", "AR3", "DR2")}

\item{starting_rr}{starting rr intervals time series value}
}
\value{
vector
}
\description{
function generating a SIMULATED rr intervals time series
}
