% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{split_all_into_runs}
\alias{split_all_into_runs}
\title{Spliting the RR time series into monotoinc runs and assigning directions}
\usage{
split_all_into_runs(rr, annotations)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}
}
\value{
a two-element list, the first element is the list of runs, the second is the vector holding the directions of the runs
}
\description{
This function splits the whole recording into disjoint monotonic runs of RR
intervals of sinus origin and assigns a direction (Up, Down, no_Change)
to each of them
}

