% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{split_on_annot}
\alias{split_on_annot}
\title{Spliting the rr time series into disjoint subseries, breaking on annotations}
\usage{
split_on_annot(rr, annotations)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}
}
\value{
a list of vectors corresponding to disjoint RR time-series of sinus origin
}
\description{
This function returns a list of disjoing segments - mini RR-time series which
only contain beats of sinus origin. The function breaks the original RR time
series on annotations which are not 0
}
