% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{time_based_jump}
\alias{time_based_jump}
\title{Function dividing the dataframe into JUMPING windows of `window` length based on time
test with lapply(ala, function(elem) sum(elem$RR) / 60 / 1000)}
\usage{
time_based_jump(RR, window = 5, cut_end = FALSE, now = "2020-09-05 12:11:00")
}
\arguments{
\item{RR}{RR data}

\item{window}{the length of the window in minutes}

\item{cut_end}{if the window does not fit the entire recording perfectly, should I cut_end = TRUE or the beginning (FALSE) of the recording}

\item{now}{when does the recording begin? default is "2020-09-05 12:11:00"}

\item{tolerance}{what departure from the window length do we tolerate}
}
\description{
Function dividing the dataframe into JUMPING windows of `window` length based on time
test with lapply(ala, function(elem) sum(elem$RR) / 60 / 1000)
}
