% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{trim_trailing_zeros}
\alias{trim_trailing_zeros}
\title{Function to trim trailing zeros}
\usage{
trim_trailing_zeros(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
A vector with trimmed trailing zeros
}
\description{
Function to trim trailing zeros
}
