% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{hrvhra}
\alias{hrvhra}
\title{HRV descriptors}
\usage{
hrvhra(
  rr,
  annotations,
  pnnX_vec = c(),
  pnn_perc_vec = c(),
  pnnX_asym = FALSE,
  pnn_perc_asym = FALSE,
  pnnX_asym_dec = TRUE,
  pnn_perc_asym_dec = TRUE,
  casethrowError = FALSE
)
}
\arguments{
\item{rr}{vector containing RR intervals time series}

\item{annotations}{vector containing annotations for the RR intervals}

\item{pnnX_asym}{whether to use the asymmetric case}

\item{pnn_perc_asym}{whether to use the asymmetric case}

\item{pnnX_asym_dec}{should the calculations be decelerations oriented}

\item{pnn_perc_asym_dec}{should the calculations be decelerations oriented}

\item{pnnX}{vector with values for calculating pnnX parameters}

\item{pnn_perc}{vector with values for calculating pnnX parameters}
}
\value{
an 1 x 10 vector containing SDNN, SD1, SD2, SD1I, SDNNd, SDNNa, SD1d, SD1a, SD2d, SD2a  descriptors
}
\description{
This function returns the basic HRV descriptors related to the Poincare plot.
}
\examples{
hrvhra(RR$RR, RR$flags)

}
\references{
J Piskorski, P Guzik, Geometry of the Poincare plot of RR intervals and its asymmetry in healthy adults, Physiological measurement 28 (3), 287 (2007)
}
