% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{index_based_slide}
\alias{index_based_slide}
\title{Function dividing the dataframe into SLIDING windows of `window` length
test with lapply(ala, function(elem) nrow(elem$RR))}
\usage{
index_based_slide(RR, window = 300, time_unit = NULL, tolerance = NULL)
}
\arguments{
\item{RR}{RR data}

\item{window}{the length of the window in beats (samples)}

\item{time_unit}{unused, dud to simplify HRAExplorer}

\item{tolerance}{unused, dud to simplify HRAExplorer}
}
\description{
Function dividing the dataframe into SLIDING windows of `window` length
test with lapply(ala, function(elem) nrow(elem$RR))
}
