% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{pnnX}
\alias{pnnX}
\title{Function calculating pnnX}
\usage{
pnnX(pp, threshold, asym = FALSE, dec = TRUE)
}
\arguments{
\item{pp}{Poincare plot object}

\item{threshold}{threshold for calculating pnnX}

\item{asym}{whether to calculate the asymmetric pnnX}

\item{dec}{whether in asymmetric pnnX we count decelerations to be bigger, false means accelerations are expected to be smaller}
}
\description{
Function calculating pnnX
}
