% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvhra.R
\name{pnnX}
\alias{pnnX}
\title{Function calculating pnnX}
\usage{
pnnX(rr, annotations, threshold)
}
\arguments{
\item{rr}{rr intervals time series}

\item{annotations}{annotations}
}
\description{
Function calculating pnnX
}
