% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{time_based_slide}
\alias{time_based_slide}
\title{Function dividing the dataframe into SLIDING windows of `window` length
test with lapply(ala, function(elem) sum(elem$RR) / 60 / 1000)}
\usage{
time_based_slide(RR, window = 5, now = "2020-09-05 12:11:00")
}
\arguments{
\item{RR}{RR data}

\item{window}{the length of the window in minutes}
}
\description{
Function dividing the dataframe into SLIDING windows of `window` length
test with lapply(ala, function(elem) sum(elem$RR) / 60 / 1000)
}
